pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.3",
    "nltk>=3.9.1",
    "pandas>=2.2.3",
    "pdfminer-six>=20250506",
    "streamlit>=1.45.0",
]
